+++ src/net/sourceforge/kolmafia/request/coinmaster/shop/CosmicRaysBazaarRequest.java
@@ -36,15 +36,14 @@ public abstract class CosmicRaysBazaarRequest extends CoinMasterShopRequest {
   // Manually set up the map and change the currency, as need
   static {
     for (Entry<Integer, Integer> entry : CoinmastersDatabase.getBuyPrices(master).entrySet()) {
-      int itemId = entry.getKey().intValue();
-      int price = entry.getValue().intValue();
+      int itemId = entry.getKey();
+      int price = entry.getValue();
       AdventureResult cost =
           switch (itemId) {
             default -> RARE_MEAT_ISOTOPE.getInstance(price);
             case ItemPool.DIGITAL_KEY -> WHITE_PIXEL.getInstance(price);
-            case ItemPool.BORIS_KEY,
-                ItemPool.JARLSBERG_KEY,
-                ItemPool.SNEAKY_PETE_KEY -> FAT_LOOT_TOKEN.getInstance(price);
+            case ItemPool.BORIS_KEY, ItemPool.JARLSBERG_KEY, ItemPool.SNEAKY_PETE_KEY ->
+                FAT_LOOT_TOKEN.getInstance(price);
             case ItemPool.RARE_MEAT_ISOTOPE -> CoinmasterData.MEAT.getInstance(price);
           };
       buyCosts.put(itemId, cost);
